requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
httpx==0.28.1
pydantic==2.10.4
pydantic-settings==2.7.0
python-multipart==0.0.20
pypdf>=5.0,<6
python-docx>=1.1.0
pymupdf>=1.24.0
